pyproject.toml
[tool.poetry]
name = "dictionarycreator"
version = "0.1.0"
description = ""
authors = ["AiKar <ira.karaseva.20@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ebooklib = "^0.17.1"
beautifulsoup4 = "^4.11.1"
pyyaml = "^6.0"
tqdm = "^4.64.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.6.0"
bandit = "^1.7.4"
mypy = "^0.991"
pip-review = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
